Reorganize release tests
